+++ doc/news.rst
@@ -9,6 +9,8 @@ News / Changelog
 (2020-XX-XX)
 
 - Sync owners and moderators from Mailman Core for MailingList. (Fixes #302)
+- Implemented a new ``HYPERKITTY_JOBS_UPDATE_INDEX_LOCK_LIFE`` setting to set
+  the lock lifetime for the ``update_and_clean_index`` job.  (Closes #300)
 
 
 1.3.3